deployer: More separation of unit and integration tests
