+++ deployer/src/lib.rs
@@ -12,10 +12,10 @@ use tracing::{error, info};
 use ulid::Ulid;
 
 mod args;
-mod deployment;
+pub mod deployment;
 mod error;
 pub mod handlers;
-mod persistence;
+pub mod persistence;
 mod proxy;
 mod runtime_manager;
